Configure gateway via environment variable (#499)
Proposed added behavior / feature: if a config file is not detected at
runtime or errors, then it'll try reading the GATEWAY_CONFIG environment
variable, which is a YAML string of the expected format. If both of
these fail the program will crash. I added a test which also functions
as a sanity test for this code

---------

Co-authored-by: Otto V <ottoevargas@gmail.com>